Backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
pymysql==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
langchain-openai==0.1.0
langchain-core==0.1.33
pydantic==2.5.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
bcrypt==4.1.2

